src/tests: Disable resetGeneration increment during ctap2 reset
This means credentials that are not discoverable won't be properly
revocated anymore. Now not that due to the fact this resetGeneration
counter was in NVM, it was reset to 0 after each app reinstallation (due
to an app update, a firmware update, or just a user triggered uninstall
then reinstall flow), and this reste was causing even more issues.